destructure the pages and components, pass the props accordingly.
