+++ src/App.jsx
@@ -4,8 +4,8 @@ import Header from "./components/Header/Header";
 // import Footer from "./components/Footer/Footer";
 import HomePage from "./pages/HomePage";
 import LoginPage from './pages/LoginPage';
-import UserProfile from './components/UserProfile/UserProfile';
-import KindAct from './components/KindAct/KindAct';
+import UserProfile from './pages/UserProfilePage';
+import KindActs from './components/KindActs/KindActs';
 // import NotFoundPage from "./pages/NotFoundPage";
 
 
@@ -17,7 +17,7 @@ function App() {
                 <Route path="/" element={<HomePage />}></Route>
                 <Route path="/login" element={<LoginPage />}></Route>
                 <Route path="/profile" element={<UserProfile />}> </Route>
-                <Route path="/kindness" element={<KindAct />}> </Route>
+                <Route path="/kindness" element={<KindActs />}> </Route>
                 {/* <Route path="*" element={<NotFoundPage />}></Route> */}
             </Routes>
             {/* <Footer /> */}
