+++ src/components/KindAct/KindAct.jsx
@@ -1,94 +1,19 @@
-import "./KindAct.scss";
-import { useState, useEffect } from "react";
-import axios from "axios";
-import { useParams } from "react-router-dom";
-import Slider from "react-slick";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
+import React from 'react'
 
-function KindAct() {
-    const api = "http://localhost:8000";
-
-    const [kindActs, setKindActs] = useState([]); // all acts of kindness
-    const [kindAct, setKindAct] = useState(null);
-    const { id } = useParams();
-
-    // Define useEffect hook to get all videos from the API
-    useEffect(() => {
-        getKindActs();
-    }, []);
-
-    // Define useEffect hook to get the selected act of kindness when the kindness id changes
-    useEffect(() => {
-        if (id) {
-            getKindAct(id);
-        } else if (kindActs.length) {
-            getKindAct(kindActs[0].id);
-        }
-    }, [id, kindActs]);
-
-    // Define function to get a selected act of kindness from the API by id
-    function getKindAct(id) {
-        axios
-            .get(`${api}/kindness/${id}`)
-            .then((res) => {
-                setKindAct(res.data);
-            })
-            .catch((err) => {
-                console.log("err: ", err);
-            })
-    };
-
-    // Define function to get all the acts of kindness from the API
-    function getKindActs() {
-        axios
-            .get(`${api}/kindness`)
-            .then((response) => {
-                setKindActs(response.data);
-                console.log(response.data)
-            })
-            .catch((err) => {
-                console.log("err: ", err);
-            })
-    };
-
-
-    const settings = {
-        dots: true,
-        infinite: true,
-        speed: 400,
-        slidesToShow: 1,
-        slidesToScroll: 1
-    }
-
-
+function KindAct({ kindActImage, kindActDescription, kindActTitle }) {
     return (
-        <section className="kindness">
-            <h2 className="kindness__header"> Choose your daily act of kindness </h2>
+        <article className="kindness__act">
+            <div className="kindness__act-wrapper">
+                <div className="kindness__text">
+                    <h3 className="kindness__title">{kindActTitle}</h3>
+                    <p className="kindness__description"> {kindActDescription}</p>
+                </div>
+                <img className="kindness__img" src={kindActImage} />
+            </div>
 
-            <Slider {...settings}>
-                {kindActs
-                    .map((kindAct) => (
-                        <>
-                            <article className="kindness__act">
-                                <div className="kindness__act-wrapper">
-                                    <div className="kindness__text">
-                                        <h3 className="kindness__title">{kindAct.title}</h3>
-                                        <p className="kindness__description"> {kindAct.description}</p>
-                                    </div>
-                                    <img className="kindness__img" src={kindAct.image} />
-                                </div>
-
-                                <button className="kindness__btn">Select</button>
-                            </article>
-
-                        </>
-                    ))}
-            </Slider>
-        </section>
-    );
+            <button className="kindness__btn">Select</button>
+        </article>
+    )
 }
-
-
 
 export default KindAct;
