+++ src/components/UserProfile/UserProfile.scss
@@ -1,0 +1,30 @@
+@use "../../styles/partials/variables" as *;
+@use "../../styles/partials/mixins" as *;
+
+.user {
+
+    &__profile {
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+        justify-content: center;
+        padding: 1rem;
+    }
+
+    &__img {
+        width: 230px;
+        text-align: center;
+        border-radius: 50%;
+    }
+
+    &__name {
+        text-align: center;
+    }
+
+    &__bio {
+        width: 40%;
+        display: flex;
+        flex-direction: column;
+        gap: 1rem;
+    }
+}
